pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cam-align"
version = "0.1.0"
description = "Portable manual camera alignment and offset compensation tool"
requires-python = ">=3.10"
dependencies = [
  "PySide6>=6.5",
  "numpy>=1.23",
  "pandas>=1.5",
  "opencv-python>=4.8",
  "tables>=3.8",
  "PyYAML>=6.0",
]

[project.scripts]
cam-align = "cam_align_tool.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
